backend_api/RAG/requirements.txt
qdrant-client
sentence-transformers
python-dotenv
pypdf
